Don't send alt unless we really mean it.
Currently Alt is pressed and released when it's activating a binding. Applications like
Firefox and all Microsoft applications (Noticed in a VM) show the application
menu when a Alt press and release is sent without another key in between. This
means that currently all M-* bindings are pretty much useless in these
applications.

This commit make's sure that we only send ALT when it's not part of a binding
that sends another binding that does not contain ALT. I also added a check that
ensure that if ALT is physically pressed and released without another key
in between the press and release, then the ALT events are sent.