refactor: :zap: Methods related to xlib, aggregated into a single class.
The plan is to reduce Display calls, which used to happen every time you searched for wm_class and wm_name, now the connection is made at the beginning and only returns information when requested.  This change also allows to return the resolution and current status of the window in relation to being in fullscreen or not.